funcion para eliminar un lugar turistico corregida

+++ src/controllers/placesController.js
@@ -9,13 +9,12 @@ exports.getAllData = async (req, res) => {
     try {
         const places = await Places.find()
         res.json(places)
-    }
 
-    catch (err) {
+    } catch (err) {
         res.status(500).json(
             { 
                 message: 'Error al obtener los lugares' 
-            });
+      	});
     }
 }
 
@@ -41,21 +40,35 @@ exports.addNewPlace = async (req, res) => {
 }
 
 
+exports.updatePLace = async (req, res) => {
+
+	const getPlaceById = re.params.id;
+	const data = Places.updateOne(getPlaceById)
+
+
+}
 
 exports.deletePlace = async (req, res) => {
 
-    try {
+		try {
         const getplaceId = await Places.findById(req.params.id)
-        getplaceId.deleteOne()
-        res.json({
-            "Message": "Place deleted!"
-        })
-    } catch (err) {
-        res.status(404).json(
-            {
-                "Message": "Place not found"
-            }
-        )
-    }
 
+				if (!getplaceId) {
+						return res.status(404).josn({
+							"message": "Error, place not found!"
+						})
+
+				} else {
+					await getplaceId.deleteOne()
+
+					res.status(200).josn({
+						"message": "Place deleted successfully!"
+					})
+				}
+        
+    } catch (err) {
+        res.status(404).json({
+              "message": "An error occurred while deleting the place"
+        })
+    }
 }
